pyproject.toml
[project]
name = "jvquant-realtime-system"
version = "0.1.0"
description = "jvQuant实时技术指标计算系统"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.28.0",
    "websocket-client>=1.4.0",
    "aiohttp>=3.8.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "mplfinance>=0.12.0",
    "python-dateutil>=2.8.0",
    "nautilus-trader @ file:///C:/Users/hedz/Downloads/nautilus_trader-develop",
    "seaborn>=0.13.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
] 
